// hc_web_angular-Develop/src/app/modules/kardex/pages/paginaPrincipal/interfaces/IlistaPacientes.ts
export interface IlistaPacientes {

    numeroDocumento: number;
    numeroIngreso: number;
    tipoDocumento: string;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    habitacion: string;
    fechaIngreso: string;
    seccion: string;
    mail: string;

}