// hc_web_angular-Develop/src/app/modules/kardex/pages/paginaPrincipal/interfaces/IlistaPacientesKar.ts
export interface IListaPacientesKar {
    seccion: string;
    descripcionSeccion: string;
    habitacion: string;
    ingreso: number;
    tipoIdentificacion: string;
    numeroIdentificacion: number;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;

}